+++ clients/privacy-center/types/api/models/MaskingStrategyConfigurationDescription.ts
@@ -1,0 +1,12 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * The description model for a specific configuration in a masking strategy
+ */
+export type MaskingStrategyConfigurationDescription = {
+  key: string;
+  optional?: boolean;
+  description: string;
+};