+++ clients/privacy-center/types/api/models/MessagingServiceDetailsMailgun.ts
@@ -1,0 +1,12 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * The details required to represent a Mailgun email configuration.
+ */
+export type MessagingServiceDetailsMailgun = {
+  is_eu_domain?: boolean | null;
+  api_version?: string | null;
+  domain: string;
+};