+++ clients/privacy-center/types/api/models/MessagingTemplateWithPropertiesBodyParams.ts
@@ -1,0 +1,9 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type MessagingTemplateWithPropertiesBodyParams = {
+  content: any;
+  properties?: Array<string> | null;
+  is_enabled: boolean;
+};