+++ clients/privacy-center/types/api/models/Page_DatasetConfigSchema_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { DatasetConfigSchema } from "./DatasetConfigSchema";
+
+export type Page_DatasetConfigSchema_ = {
+  items: Array<DatasetConfigSchema>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};