+++ clients/privacy-center/types/api/models/SupportedLanguage.ts
@@ -2,9 +2,6 @@
 /* tslint:disable */
 /* eslint-disable */
 
-/**
- * An enumeration.
- */
 export enum SupportedLanguage {
   AR = "ar",
   BG = "bg",
@@ -16,6 +13,7 @@ export enum SupportedLanguage {
   EL = "el",
   EN = "en",
   ES = "es",
+  ES_MX = "es-MX",
   ET = "et",
   EU = "eu",
   FI = "fi",