+++ clients/privacy-center/types/api/models/TCFPurposeOverrideSchema.ts
@@ -1,0 +1,14 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { TCFLegalBasisEnum } from "./TCFLegalBasisEnum";
+
+/**
+ * TCF Purpose Override Schema
+ */
+export type TCFPurposeOverrideSchema = {
+  purpose: number;
+  is_included?: boolean | null;
+  required_legal_basis?: TCFLegalBasisEnum | null;
+};