+++ clients/privacy-center/types/api/models/ValidationStatus.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * Validate endpoint response status
+ */
+export enum ValidationStatus {
+  SUCCESS = "success",
+  FAILURE = "failure",
+}